tl;dr shit works, yo
Specifically, I've started using the nonce that gets POST'd to the server in the payment.
Test card 4111 1111 1111 1111 should always succeed.

I'm redirecting to a success page for successful payments.
Not redirecting to a failure page yet, 'cause i wanna test that, and it's easier if it doesn't redirect.

I'm also logging (quickly and dirtily with warn):
- the nonce
- the result (and result.inspect)
- whether it was successful