Basic braintree error handling.
I've tried a few of the test nonces from this url
https://developers.braintreepayments.com/javascript+ruby/start/hello-server#test-values
Got a good idea of how the errors work.

For BattleHack, we don't really care much beyond whether it fails or succeeds.
We'll only really be demoing the happy path, but it's useful just in case something does go wrong.

I've also enabled session variables, 'cause those are useful.